Add separated controller for StoreTrafficFlows handler
