+++ app/api/controllers.go
@@ -39,19 +39,6 @@ func (api *Api) GetActiveFlowsPerDestination(c *gin.Context) {
 	c.JSON(http.StatusOK, gin.H{"data": activeFlows})
 }
 
-func (api *Api) StoreActiveTraffic(c *gin.Context) {
-	err := api.StoreTrafficFlowsUseCase.StoreTrafficFlows()
-	if err != nil {
-		fmt.Println(err)
-		c.JSON(500, gin.H{"data": "error"})
-		c.AbortWithStatus(http.StatusInternalServerError)
-		return
-	}
-	c.Header("Access-Control-Allow-Origin", "*") //There is a vuln here, that's only for testing purpose.
-	c.Header("Access-Control-Allow-Methods", "POST")
-	c.JSON(http.StatusOK, gin.H{"message": "ok"})
-}
-
 func (api *Api) GetActiveFlowsPerCountry(c *gin.Context) {
 	activeFlows, err := api.TrafficBytesParser.GetBytesPerCountry()
 	if err != nil {
